ZOOKEEPER-2775: ZK Client not able to connect with Xid out of order error
Once client enters into Xid out of order issue, It never comes to normal state. It keeps trying to connect and  fail with the same error.  Recreating/Restarting is the only solution as of now. This happens because of bug in the ZK client code. This MR provides the fix.

Author: Mohammad Arshad <arshad@apache.org>

Reviewers: Michael Han <hanm@apache.org>, Rakesh Radhakrishnan <rakeshr@apache.org>, Abe Fine <afine@apache.org>

Closes #254 from arshadmohammad/ZOOKEEPER-2775-XidOutOfOrder