ZOOKEEPER-3404: Downgrade BouncyCastle to 1.60
I've seen a lot of test timeout errors with QuorumSSL tests since I upgraded master to BouncyCastle 1.61 due to a Java 9 warning. The warning has been reported by Enrico Olivelli which we tried to solve by the upgrade, but the warning message is still present so I don't see any harm in downgrading to the previous version.

https://issues.apache.org/jira/browse/ZOOKEEPER-3404

Author: Andor Molnar <andor@apache.org>

Reviewers: Enrico Olivelli <eolivelli@apache.org>, Norbert Kalmar <nkalmar@apache.org>

Closes #961 from anmolnar/ZOOKEEPER-3404