ZOOKEEPER-3455: fix UnifiedServerSocketTest on jdk 13
The `UnifiedServerSocketTest.testConnectWithoutSSLToStrictServer` fails on OpenJDK 13 because in the new default socket implementation (https://openjdk.java.net/jeps/353) the `NioSocketImpl.getInputStream.read()` behaves differently than the old `SocketInputStream.read()`.

A workaround could be to execute the tests using the `-Djdk.net.usePlainSocketImpl` system property (hardcoding it in the maven / ant configs), which enforces the usage of the old socket implementation in JDK 13. But I preferred instead to make the test compatible with both older and newer JDK, so it should succeed even if someone is executing it outside of our build environment.

Author: Mate Szalay-Beko <szalay.beko.mate@gmail.com>

Reviewers: andor@apache.org

Closes #1029 from symat/master