ZOOKEEPER-3981: Flaky test MultipleAddressTest::testGetValidAddressWithNotValid
Problem:

Test MultipleAddressTest::testGetValidAddressWithNotValid might fail deterministically when the address it's using, 10.0.0.1, is reachable, as per https://tools.ietf.org/html/rfc5735 10.0.0.1 might be allocatable to private network usage. In fact, the router address of my ISP is assigned this IP, leading to this test always failing for me.

Solution:

Replace the address with 240.0.0.0, which is reserved for future use and less likely to be reachable.

Author: Michael Han <michael@lianghan.org>
Author: Michael Han <hanm@apache.org>

Reviewers: Christopher Tubbs <ctubbsii@apache.org>, Mate Szalay-Beko <symat@apache.org>

Closes #1511 from hanm/fixut