ZOOKEEPER-3874: Official API to start ZooKeeper server from Java
Create an Official API to start a ZooKeeper server node from Java code.
The idea is not to run ZooKeeper server inside the same process of an application, but only to have a standard Launcher that can be used from Java and not a bash script.

See more context here
https://issues.apache.org/jira/browse/ZOOKEEPER-3874

This is how a Java launcher will look like for tests:
```
        int clientPort = PortAssignment.unique();
        final Properties configZookeeper = new Properties();
        configZookeeper.put("clientPort", clientPort + "");
        configZookeeper.put("host", "localhost");
        configZookeeper.put("..........................");
        try (ZooKeeperServerEmbedded zkServer = ZooKeeperServerEmbedded
                .builder()
                .baseDir(baseDir)
                .configuration(configZookeeper)
                .exitHandler(ExitHandler.LOG_ONLY)
                .build()) {
            zkServer.start();
            //// wait.....
      }
```

This feature does not overlap with Curator TestingServer, this feature is meant to be used a fundation for projects like TestingServer but also to run ZooKeeper server nodes in production.

This code is running in production at https://www.mag-news.com and https://emailsuccess.com, in such products we are using a Java based process manager

Author: Enrico Olivelli <eolivelli@apache.org>
Author: Enrico Olivelli <eolivelli@gmail.com>

Reviewers: Damien Diederen <ddiederen@apache.org>

Closes #1526 from eolivelli/fix/ZOOKEEPER-3874-embedded-api