ZOOKEEPER-4030: Optionally canonicalize host names in quorum SASL authentication
Without the option provided by this changeset, it is not convenient to use `CNAME`s to reference servers in `server.x` configuration lines while using quorum Kerberos authentication, as "incorrect" principals are generated by default.

Setting `kerberos.canonicalizeHostNames=true` causes the quorum server to substitute the canonical host name for the one specified in the configuration.  This has an effect on principal generation, and on the filtering of authorization IDs from incoming connections.

Author: Damien Diederen <dd@crosstwine.com>

Reviewers: Enrico Olivelli <eolivelli@apache.org>

Closes #1564 from ztzg/ZOOKEEPER-4030-quorum-sasl-auth-with-cnames